Working tests, lint, types
