+++ src/contexts/FilterContext.tsx
@@ -1,15 +1,6 @@
-import React, { createContext, useState, useContext, ReactNode } from 'react';
+import React, { useState, ReactNode } from 'react';
 import { FilterModel } from 'ag-grid-community';
-
-interface FilterContextType {
-  filterModel: FilterModel | null;
-  setFilterModel: (model: FilterModel | null) => void;
-  visibleTransactionIds: string[];
-  setVisibleTransactionIds: (ids: string[]) => void;
-  resetFilters: () => void;
-}
-
-const FilterContext = createContext<FilterContextType | undefined>(undefined);
+import { FilterContext } from './FilterContextValue';
 
 export const FilterProvider: React.FC<{ children: ReactNode }> = ({ children }) => {
   const [filterModel, setFilterModel] = useState<FilterModel | null>(null);
@@ -38,12 +29,4 @@ export const FilterProvider: React.FC<{ children: ReactNode }> = ({ children }) 
       {children}
     </FilterContext.Provider>
   );
-};
-
-export const useFilterContext = () => {
-  const context = useContext(FilterContext);
-  if (context === undefined) {
-    throw new Error('useFilterContext must be used within a FilterProvider');
-  }
-  return context;
 };
