+++ src/contexts/useFilterContext.tsx
@@ -1,0 +1,10 @@
+import { useContext } from 'react';
+import { FilterContext } from './FilterContextValue';
+
+export const useFilterContext = () => {
+  const context = useContext(FilterContext);
+  if (context === undefined) {
+    throw new Error('useFilterContext must be used within a FilterProvider');
+  }
+  return context;
+};
